start on the arrays, dict, stream, off to the side
